Version change from 0.2.2 to 0.3
Changes from Jason: added the possibility to explicitly read in files as labels. These files will be stored with a valid image range spanning the values present in the labels in order to avoid any slice scaling. The argument "labels" (can be set to True or False) are added to all factory functions: volumeFromFile, volumeFromInstance, volumeLikeFile, volumeFromDescription.

In volumes.py, in the function loadData: if a volumeFromInstance was created without copying the data from the instance, the numpy data array would always be created as a float array (default from the zeros function). Changed that to include the dtype that was giving to volumeFromInstance.

Added history information to files that are create/modified by pyminc. When volumeFromDescription is invoked, a new history entry is created. In all other cases, the history from the like file is copied. In order to ensure backwards compatibility, if the program calling pyminc function does not append the history itself, a warning is displayed and the line "modified using pyminc" is added to the history. In general, before calling writeFile(), the function appendAndWriteHistory(history="string with info") should be called.

Changes related to the current version of libminc (GitHub: commit 3ad994c241dd027):
1) in miclose_volume, the memory for the dimensions is freed already, so I removed the lines that freed this memory in the closeVolume function (led to a seg fault)
2) type changes: a number of integer variables have changed to misize_t type (= hsize_t = unsigned long long). These changes occur in the following functions: miget_dimension_sizes, miget_dimension_separations, miget_real_value_hyperslab, miset_real_value_hyperslab, micreate_dimension, miadd_history_attr, miget_attr_values

Bug fixes: in mincSizes["uint"]... changed unit32 to uint32
Bug fixes: in the function loadData, removed the line " self.dtype = dtype". The variable dtype is not around in that function anymore, and self.dtype is properly set already. This line caused the dtype of any object to be set to "numpy.dtype".